// frontend/lib/api/generated/model/requerimientoAnualDetailOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { RequerimientoAnualDetalleOut } from './requerimientoAnualDetalleOut';

export interface RequerimientoAnualDetailOut {
  requerimiento_anual_id: number;
  racion_anual_id: number;
  version: number;
  estado: string;
  presupuesto_referencial_total: number | null;
  aprobado_por_id: number | null;
  aprobado_en: string | null;
  creado_en: string;
  detalle: RequerimientoAnualDetalleOut[];
}